🐛 Fix real number types

+++ coursework/entity-jewel.h
@@ -7,7 +7,7 @@ struct Jewel {
     double weight;
     int number;
     char stone_name[50];
-    int carat_weight;
+    double carat_weight;
     double price;
     char date[11];
     struct Jewel* next;
